Move canAddDisplay and addDisplayToSettings to own js module
Move both functions and their tests to own js modules.
