+++ gsa/src/web/pages/start/dashboard.js
@@ -24,7 +24,7 @@ import React from 'react';
 
 import PropTypes from 'web/utils/proptypes';
 
-import {canAddDisplay} from 'web/store/dashboard/settings/actions';
+import {canAddDisplay} from 'web/store/dashboard/settings/utils';
 
 import {Dashboard} from 'web/components/dashboard/dashboard';
 import {DashboardControls} from 'web/components/dashboard/controls';
