+++ gsa/src/web/store/dashboard/settings/__tests__/actions.js
@@ -25,8 +25,6 @@
 
 import {isFunction} from 'gmp/utils/identity';
 
-import {DEFAULT_ROW_HEIGHT} from 'web/components/sortable/utils';
-
 import {
   DASHBOARD_SETTINGS_LOADING_ERROR,
   DASHBOARD_SETTINGS_LOADING_REQUEST,
@@ -43,8 +41,6 @@ import {
   loadSettings,
   saveSettings,
   resetSettings,
-  canAddDisplay,
-  addDisplayToSettings,
 } from '../actions';
 
 const createRootState = (state = {byId: {}}) => ({
@@ -122,200 +118,6 @@ describe('dashboard settings action tests', () => {
 
 });
 
-describe('canAddDisplay helper function tests', () => {
-
-  test('should allow to add display without settings', () => {
-    expect(canAddDisplay()).toEqual(true);
-  });
-
-  test('should allow to add display without existing rows', () => {
-    const rows = [];
-    expect(canAddDisplay({
-      rows,
-      maxItemsPerRow: 3,
-      maxRows: 1,
-    })).toEqual(true);
-  });
-
-  test('should allow to add display without maxItemsPerRow', () => {
-    const rows = [
-      {
-        items: [1, 2, 3],
-      },
-    ];
-    expect(canAddDisplay({
-      rows,
-      maxRows: 1,
-    })).toEqual(true);
-  });
-
-  test('should allow to add display without maxRows', () => {
-    const rows = [
-      {
-        items: [1, 2, 3],
-      },
-    ];
-    expect(canAddDisplay({
-      rows,
-      maxItemsPerRow: 3,
-    })).toEqual(true);
-  });
-
-  test('should allow to add display if last row is not full', () => {
-    const rows = [
-      {
-        items: [1, 2],
-      },
-    ];
-    expect(canAddDisplay({
-      rows,
-      maxItemsPerRow: 3,
-      maxRows: 1,
-    })).toEqual(true);
-  });
-
-  test('should allow to add display if another row is possible', () => {
-    const rows = [
-      {
-        items: [1, 2, 3],
-      },
-    ];
-    expect(canAddDisplay({
-      rows,
-      maxItemsPerRow: 3,
-      maxRows: 2,
-    })).toEqual(true);
-  });
-
-  test('should not be able to add display if last row is full', () => {
-    const rows = [
-      {
-        items: [1, 2, 3],
-      },
-    ];
-    expect(canAddDisplay({
-      rows,
-      maxItemsPerRow: 3,
-      maxRows: 1,
-    })).toEqual(false);
-  });
-
-});
-
-describe('addDisplayToSettings tests', () => {
-
-  test('should create new settings', () => {
-    const uuid = 'foo1';
-    const uuidFunc = () => uuid;
-
-    expect(addDisplayToSettings(undefined, 'a1', uuidFunc)).toEqual({
-      rows: [{
-        items: [{
-          id: uuid,
-          name: 'a1',
-        }],
-      }],
-    });
-
-    expect(addDisplayToSettings({rows: {}}, 'a1', uuidFunc)).toEqual({
-      rows: [{
-        items: [{
-          id: uuid,
-          name: 'a1',
-        }],
-      }],
-    });
-
-    expect(addDisplayToSettings({rows: []}, 'a1', uuidFunc)).toEqual({
-      rows: [{
-        items: [{
-          id: uuid,
-          name: 'a1',
-        }],
-      }],
-    });
-  });
-
-  test('should add display to last row', () => {
-    const uuid = 'foo1';
-    const uuidFunc = () => uuid;
-    const settings = {
-      rows: [{
-        items: [{
-          id: 'u1',
-        }],
-      }, {
-        items: [{
-          id: 'u2',
-        }],
-      }],
-      maxItemsPerRow: 2,
-    };
-
-    const newSettings = addDisplayToSettings(settings, 'a1', uuidFunc);
-    expect(newSettings).toEqual({
-      rows: [{
-        items: [{
-          id: 'u1',
-        }],
-      }, {
-        items: [{
-          id: 'u2',
-        }, {
-          id: uuid,
-          name: 'a1',
-        }],
-      }],
-      maxItemsPerRow: 2,
-    });
-    expect(settings).not.toBe(newSettings);
-    // ensure changed row has copied instead of mutated
-    expect(settings.rows[0]).toBe(newSettings.rows[0]);
-    expect(settings.rows[0].items).toBe(newSettings.rows[0].items);
-    expect(settings.rows[1]).not.toBe(newSettings.rows[1]);
-    expect(settings.rows[1].items).not.toBe(newSettings.rows[1].items);
-  });
-
-  test('should add display to new row', () => {
-    const uuid = 'foo1';
-    const uuidFunc = () => uuid;
-    const settings = {
-      rows: [{
-        items: [{
-          id: 'u1',
-        }],
-      }, {
-        items: [{
-          id: 'u2',
-        }],
-      }],
-      maxItemsPerRow: 1,
-    };
-
-    const newSettings = addDisplayToSettings(settings, 'a1', uuidFunc);
-    expect(newSettings).toEqual({
-      rows: [{
-        items: [{
-          id: 'u1',
-        }],
-      }, {
-        items: [{
-          id: 'u2',
-        }],
-      }, {
-        id: uuid,
-        height: DEFAULT_ROW_HEIGHT,
-        items: [{
-          id: uuid,
-          name: 'a1',
-        }],
-      }],
-      maxItemsPerRow: 1,
-    });
-    expect(settings).not.toBe(newSettings);
-  });
-
-});
 
 describe('loadSettings tests', () => {
